massive refactoring
