+++ src/anemoi/datasets/create/input/empty.py
@@ -16,9 +16,9 @@ from anemoi.datasets.dates import DatesProvider as DatesProvider
 from anemoi.datasets.fields import FieldArray as FieldArray
 from anemoi.datasets.fields import NewValidDateTimeField as NewValidDateTimeField
 
-from ..trace import trace_datasource
 from .misc import assert_fieldlist
 from .result import Result
+from .trace import trace_datasource
 
 LOG = logging.getLogger(__name__)
