+++ src/anemoi/datasets/create/input/filter.py
@@ -17,13 +17,13 @@ from anemoi.datasets.fields import FieldArray as FieldArray
 from anemoi.datasets.fields import NewValidDateTimeField as NewValidDateTimeField
 
 from ..functions import import_function
-from ..template import notify_result
-from ..trace import trace_datasource
 from .function import FunctionContext
 from .misc import _tidy
 from .misc import assert_fieldlist
 from .step import StepAction
 from .step import StepResult
+from .template import notify_result
+from .trace import trace_datasource
 
 LOG = logging.getLogger(__name__)
