+++ src/anemoi/datasets/create/input/template.py
